- couple tests added during live coding - implemented new figure and refactor into -> double - refactoring in PmAlarmService

+++ 02_figures-tests/src/main/java/foo/bar/Figure.java
@@ -1,6 +1,6 @@
 package foo.bar;
 
 interface Figure {
-    int area();
+    double area();
     int circumference();
 }
